LayoutTests/editing/caret/ios/fixed-caret-position-after-scroll.html

<!DOCTYPE html> <!-- webkit-test-runner [ useFlexibleViewport=true ] -->
<html>
<meta name="viewport" content="width=device-width, initial-scale=1, user-scalable=no">
<head>
    <script src="../../../resources/js-test-pre.js"></script>
    <style>
        body {
            margin: 0;
        }

        #input {
            width: 100%;
            height: 50px;
            position: fixed;
            left: 0;
            top: 0;
        }

        div {
            background-image: linear-gradient(blue, red);
            height: 4000px;
        }
    </style>
    <script>
    if (window.testRunner) {
        testRunner.dumpAsText();
        testRunner.waitUntilDone();
    }

    function getInputViewBoundsAfterTappingInInputScript(tapX, tapY)
    {
        return `(function() {
            uiController.didShowKeyboardCallback = function() {
                uiController.doAfterNextStablePresentationUpdate(function() {
                    uiController.uiScriptComplete(JSON.stringify(uiController.inputViewBounds));
                });
            };
            uiController.singleTapAtPoint(${tapX}, ${tapY}, function() { });
        })()`;
    }

    function getCaretRectAfterScrollToOffsetScript(distance)
    {
        return `(function() {
            uiController.stableStateOverride = false;
            uiController.immediateScrollToOffset(0, ${distance});
            uiController.stableStateOverride = true;
            uiController.doAfterNextStablePresentationUpdate(function() {
                uiController.uiScriptComplete(JSON.stringify(uiController.textSelectionCaretRect));
            });
        })()`;
    }

    function toString(rect)
    {
        return `[${rect.left} ${rect.top} ; ${rect.width} ${rect.height}]`;
    }

    function run()
    {
        if (!window.testRunner || !testRunner.runUIScript) {
            description("To manually test, tap this input field and scroll up. The text caret should not end up outside of the input.");
            return;
        }

        testRunner.runUIScript(getInputViewBoundsAfterTappingInInputScript(window.innerWidth / 2, 30), (inputViewBounds) => {
            window.inputViewBounds = inputViewBounds = JSON.parse(inputViewBounds);
            testRunner.runUIScript(getCaretRectAfterScrollToOffsetScript(inputViewBounds.height), initialCaretRect => {
                window.initialCaretRect = initialCaretRect = JSON.parse(initialCaretRect);
                testRunner.runUIScript(getCaretRectAfterScrollToOffsetScript(inputViewBounds.height + 500), finalCaretRect => {
                    window.finalCaretRect = finalCaretRect = JSON.parse(finalCaretRect);
                    shouldBe("finalCaretRect.top - initialCaretRect.top", "500");
                    testRunner.notifyDone();
                });
            });
        });
    }
    </script>
</head>
<body onload=run()>
    <input id="input"></input>
    <script src="../../../resources/js-test-post.js"></script>
</body>

</html>
